requirements.txt
python-dotenv
requests
Flask
flask_login
psycopg2-binary
Flask-SQLAlchemy
waitress